backend/requirements.txt
fastapi
fastapi-mail
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary  # For NeonDB (Postgres)
pypdf            # For Resume Parsing
python-multipart # For file uploads
deepgram-sdk==3.* 
groq
websockets
httpx            # Async HTTP client